fix reviewboot.

+++ services/network_search/src/network_utils.cpp
@@ -97,7 +97,7 @@ static const int32_t ALL_RAF[] = { GSM, CDMA, EVDO, WCDMA, TDSCDMA, LTE, NR };
 
 PreferredNetworkMode NetworkUtils::GetNetworkModeFromRaf(int32_t raf)
 {
-    for (int32_t i = 0; i < sizeof(ALL_RAF) / sizeof(ALL_RAF[0]); i++) {
+    for (int32_t i = 0; i < static_cast<int32_t>(sizeof(ALL_RAF) / sizeof(ALL_RAF[0])); i++) {
         if (static_cast<uint32_t>(ALL_RAF[i]) & static_cast<uint32_t>(raf)) {
             raf = static_cast<int32_t>(static_cast<uint32_t>(ALL_RAF[i]) | static_cast<uint32_t>(raf));
         }
